+++ API/main.py
@@ -6,7 +6,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from starlette.middleware.sessions import SessionMiddleware
 
-from API import collection, item, user, wardrobe
+from API import collection, user, wardrobe
 from API.database import DB, get_database
 
 app = FastAPI()
@@ -19,7 +19,6 @@ origins = [
 ]
 
 app.include_router(collection.router, tags=["collection"])
-app.include_router(item.router, tags=["item"])
 app.include_router(wardrobe.router, tags=["wardrobe"])
 app.include_router(user.router, tags=["user"])
 app.include_router(ai_model.router, tags=["ai"])
